# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/permission/:id": {
            "post": {
                "description": "Update Permission by role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Update Permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Permission",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rules.PermissionUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/permission/{id}": {
            "get": {
                "description": "List Permission by role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "List Permission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "With the bearer started",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rules.PermissionUpdate": {
            "type": "object",
            "required": [
                "permission"
            ],
            "properties": {
                "permission": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rules.UpdateRequestPermission"
                    }
                }
            }
        },
...
        "rules.UpdateRequestPermission": {
            "type": "object",
            "required": [
                "page",
                "resource"
            ],
            "properties": {
                "page": {
                    "type": "string"
                },
                "resource": {
                    "type": "string"
                }
            }
        }
    }
}